@@ -1,41 +1,54 @@
 {
-  "name": "rml-studio",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@ducanh2912/next-pwa": "^10.2.6",
-    "@nextui-org/button": "^2.0.30",
-    "@nextui-org/react": "^2.3.5",
-    "@nextui-org/system": "^2.1.2",
-    "@nextui-org/theme": "^2.2.2",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "framer-motion": "^11.1.7",
-    "lucide-react": "^0.378.0",
-    "next": "14.2.3",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-icons": "^5.1.0",
-    "react-loader-spinner": "^6.1.6",
-    "tailwind-merge": "^2.3.0",
-    "tailwindcss-animate": "^1.0.7"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.3",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.3",
-    "typescript": "^5",
-    "webpack": "^5.91.0"
-  }
+	"name": "next-app-template",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint"
+	},
+	"dependencies": {
+		"@nextui-org/button": "^2.0.31",
+		"@nextui-org/code": "^2.0.27",
+		"@nextui-org/input": "^2.1.21",
+		"@nextui-org/kbd": "^2.0.28",
+		"@nextui-org/link": "^2.0.29",
+		"@nextui-org/navbar": "^2.0.30",
+		"@nextui-org/react": "^2.3.6",
+		"@nextui-org/snippet": "^2.0.35",
+		"@nextui-org/switch": "^2.0.28",
+		"@nextui-org/system": "2.1.0",
+		"@nextui-org/theme": "2.2.0",
+		"@radix-ui/react-slot": "^1.0.2",
+		"@react-aria/ssr": "^3.9.2",
+		"@react-aria/visually-hidden": "^3.8.10",
+		"@tailwindcss/container-queries": "^0.1.1",
+		"@types/node": "20.5.7",
+		"@types/react": "18.2.21",
+		"@types/react-dom": "18.2.7",
+		"autoprefixer": "10.4.19",
+		"class-variance-authority": "^0.7.0",
+		"clsx": "^2.1.1",
+		"currency-icons": "^1.0.14",
+		"eslint": "8.48.0",
+		"eslint-config-next": "14.2.1",
+		"framer-motion": "^11.1.1",
+		"intl-messageformat": "^10.5.0",
+		"lucide-react": "^0.378.0",
+		"next": "14.2.1",
+		"next-themes": "^0.2.1",
+		"postcss": "8.4.38",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"rippleui": "^1.12.1",
+		"tailwind-extended-shadows": "^0.4.0",
+		"tailwind-merge": "^2.3.0",
+		"tailwind-variants": "^0.1.20",
+		"tailwindcss": "3.4.3",
+		"typescript": "5.0.4"
+	},
+	"devDependencies": {
+		"tailwindcss-animate": "^1.0.7"
+	}
 }
