@@ -1,45 +1,72 @@
-import Image from "next/image";
-import { Button } from "@nextui-org/button";
-import ForeverMobileScreen from "./ForeverMobileScreen";
-import TopNavbar from "./TopNavbar";
-import AccountOverview from "./AccountOverview";
+import { Link } from "@nextui-org/link";
+import { Snippet } from "@nextui-org/snippet";
+import { Code } from "@nextui-org/code";
+import { button as buttonStyles } from "@nextui-org/theme";
+import { siteConfig } from "@/config/site";
+import { title, subtitle } from "@components/primitives";
+import { GithubIcon } from "@components/icons";
+import { Navbar } from "@components/navbar";
 
 export default function Home() {
   return (
-    <>
-      <TopNavbar />
-      <main className="light">
-        <section className="pt-12 pb-20">
-          
+    <div className="relative flex flex-col h-screen">
+      <Navbar />
+      <main className="container mx-auto max-w-7xl pt-16 px-6 flex-grow">
+        <section className="flex flex-col items-center justify-center gap-4 py-8 md:py-10">
+          <div className="inline-block max-w-lg text-center justify-center">
+            <h1 className={title()}>Make&nbsp;</h1>
+            <h1 className={title({ color: "violet" })}>beautiful&nbsp;</h1>
+            <br />
+            <h1 className={title()}>
+              websites regardless of your design experience.
+            </h1>
+            <h2 className={subtitle({ class: "mt-4" })}>
+              Beautiful, fast and modern React UI library.
+            </h2>
+          </div>
+
+          <div className="flex gap-3">
+            <Link
+              isExternal
+              href={siteConfig.links.docs}
+              className={buttonStyles({
+                color: "primary",
+                radius: "full",
+                variant: "shadow",
+              })}
+            >
+              Documentation
+            </Link>
+            <Link
+              isExternal
+              className={buttonStyles({ variant: "bordered", radius: "full" })}
+              href={siteConfig.links.github}
+            >
+              <GithubIcon size={20} />
+              GitHub
+            </Link>
+          </div>
+
+          <div className="mt-8">
+            <Snippet hideSymbol hideCopyButton variant="flat">
+              <span>
+                Get started by editing <Code color="primary">app/page.tsx</Code>
+              </span>
+            </Snippet>
+          </div>
         </section>
-        <section className="pt-8 pb-20">
-        <AccountOverview />
-      </section>
-
-      <section>
-        <h2>Quick Actions</h2>
-        <div>
-          <ul>
-            <li>
-              <button>Redeem Giftcard</button>
-            </li>
-            <li>
-              <button>Upload Music</button>
-            </li>
-            <li>
-              <button>Download Beats</button>
-            </li>
-            <li>
-              <button>Support & Donate</button>
-            </li>
-          </ul>
-        </div>
-      </section>
-
-      <section>
-        
-      </section>
       </main>
-    </>
+      <footer className="w-full flex items-center justify-center py-3">
+        <Link
+          isExternal
+          className="flex items-center gap-1 text-current"
+          href="https://nextui-docs-v2.vercel.app?utm_source=next-app-template"
+          title="nextui.org homepage"
+        >
+          <span className="text-default-600">Powered by</span>
+          <p className="text-primary">NextUI</p>
+        </Link>
+      </footer>
+    </div>
   );
 }
