@@ -1,17 +1,22 @@
-'use client'
+"use client";
 
-import {NextUIProvider} from '@nextui-org/react'
-import { IconContext } from "react-icons";
+import * as React from "react";
+import { NextUIProvider } from "@nextui-org/system";
+import { useRouter } from 'next/navigation'
+import { ThemeProvider as NextThemesProvider } from "next-themes";
+import { ThemeProviderProps } from "next-themes/dist/types";
 
-export function Providers({children}: { children: React.ReactNode }) {
-  return (
-    <NextUIProvider>
-      {/* <IconContext.Provider value={{ color: "blue", className: "global-class-name" }}> */}
-      <IconContext.Provider value={{}}>
-      {children}
-      </IconContext.Provider>
-    </NextUIProvider>
-  )
+export interface ProvidersProps {
+	children: React.ReactNode;
+	themeProps?: ThemeProviderProps;
 }
 
+export function Providers({ children, themeProps }: ProvidersProps) {
+  const router = useRouter();
 
+	return (
+		<NextUIProvider navigate={router.push}>
+			<NextThemesProvider {...themeProps}>{children}</NextThemesProvider>
+		</NextUIProvider>
+	);
+}
